step module to generate JSON and BED files added (#14)
